pyproject.toml
[project]
name = "google-api-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client==2.45.0",
    "google-auth-oauthlib>=1.2.4",
]
